Bugfix for photon selection in ATLAS_2017_I1644367
The IdentifiedFinalState without acceptId() would not select any particles, so this can't have worked like this. The ID selection was anyway already present as an explicit cut, so we don't need the IdentifiedFinalState anyway.

Since the ATLAS analysis is selecting prompt photons only, I have changed this to a PromptFinalState instead of just a FinalState.