make-plots: load specific pst modules instead of pst-all
pst-all includes various packages that we don't need:

http://tug.ctan.org/graphics/pstricks/base/latex/pst-all.sty

Loading those takes away some of the resources in tex, so some bigger
histograms may fail to render with:

  ! TeX capacity exceeded

This does not completely remove the problem, but raises the bar for
which histograms fail. This was tested to work with basic graphics
options: LineStyle, FillColor, FillOpacity, HatchColor.