Fix direction of ALICE V0 Multiplicity projections.
- V0A is in the _positive_ pseudorapidity direction
- V0C is in the _negative_ pseudorapidity direction

That means for V0A calibrations to make sense for p-Pb collisions, the
data _should_ have that the lead ion has positive rapidity (projectile)
and the  proton _should_ have negative rapidity (target).

This should probably be noted in the `.info` of the calibration
analyses `ALICE_2015_PPB_CENTRALITY` and analyses that uses the
V0A centrality estimator.

Closes #189