rivet-cmphistos: fix for PLOT: argument in any position
There is a bug in rivet-cmphistos which messes up the plots when the
PLOT: options are specified, but not as a last argument.

rivet-mkhtml PLOT:LogY=0 output.yoda.gz:title

The logic in the rivet-cmphistos relies on zipping filelist and
filenames. However, the "PLOT" value is only being removed only from the
former. This change solves the problem by not adding "PLOT" values to
either of those lists in the first place.